+++ src/tools/mod.rs
@@ -1,4 +1,3 @@
-
-pub(crate) mod foc_pid;
 pub(crate) mod foc_command;
+pub(crate) mod foc_pid;
 pub(crate) mod shaft_position;