+++ release_notes.md
@@ -10,3 +10,10 @@
 * Updated some comments in code, and command help menu
 * Updated README.md
 * Made a comment in code  that part of the code in foc_pwm is copied from library foc v0.3.0
+
+## Version 0.1.2 dd 2025-09-09
+
+* Calibration can be done at any time with about same result
+* Torque used for calibration is default 0.25, but can be set before calibration with the tt command
+* bugfix for calibration routine
+* Added example how to measure the pwm output of an angle sensor